test(core): split watch sourcemap tests
