+++ packages/core/test/core/watch/source-maps-basic/bundle-getters/disabled.ts
@@ -1,0 +1,33 @@
+import {
+  describe,
+  expect,
+  join,
+  mkdtempSync,
+  rmSync,
+  test,
+  tmpdir,
+  watch,
+  writeFileSync,
+} from '../helpers';
+
+describe('watch() > source maps basic > bundle getters > disabled', () => {
+  test('getBundleSourceMap — sourcemap 비활성 시 null', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-lazy-sm-off-'));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 1;\n');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    const handle = watch({
+      entryPoints: [join(dir, 'entry.ts')],
+      outfile: join(dir, 'bundle.js'),
+      devMode: true,
+      onReady() {
+        readyDone();
+      },
+    });
+    await readyP;
+
+    expect(handle.getBundleSourceMap()).toBeNull();
+    handle.stop();
+    rmSync(dir, { recursive: true });
+  }, 10000);
+});
