+++ packages/core/test/core/watch/source-maps-basic/hmr-getters/initial-build.ts
@@ -1,0 +1,50 @@
+import {
+  describe,
+  expect,
+  join,
+  mkdtempSync,
+  rmSync,
+  test,
+  tmpdir,
+  watch,
+  writeFileSync,
+} from '../helpers';
+
+describe('watch() > source maps basic > hmr getters > initial build', () => {
+  test('getHmrSourceMap — initial build 직후 (rebuild 전) 모듈 id 조회 가능', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-lazy-init-'));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 1;\n');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    const { promise: rebuildP, resolve: rebuildDone } = Promise.withResolvers<{
+      updates?: Array<{ id: string }>;
+    }>();
+    const handle = watch({
+      entryPoints: [join(dir, 'entry.ts')],
+      outfile: join(dir, 'bundle.js'),
+      sourcemap: true,
+      devMode: true,
+      emitDiskSourcemap: false,
+      onReady() {
+        readyDone();
+      },
+      onRebuild(event) {
+        rebuildDone(event);
+      },
+    });
+    await readyP;
+
+    await new Promise((r) => setTimeout(r, 100));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 2;\n');
+    const event = await rebuildP;
+    const id = event.updates![0].id;
+
+    const json = handle.getHmrSourceMap(id);
+    expect(json).not.toBeNull();
+    const m = JSON.parse(json!);
+    expect(m.version).toBe(3);
+
+    handle.stop();
+    rmSync(dir, { recursive: true });
+  }, 15000);
+});
