+++ packages/core/test/core/watch/source-maps-basic/hmr-getters/module-id.ts
@@ -1,0 +1,53 @@
+import {
+  describe,
+  expect,
+  join,
+  mkdtempSync,
+  rmSync,
+  test,
+  tmpdir,
+  watch,
+  writeFileSync,
+} from '../helpers';
+
+describe('watch() > source maps basic > hmr getters > module id', () => {
+  test('getHmrSourceMap — 모듈 id 로 JSON 반환, 미존재 id 는 null', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-lazy-hmr-sm-'));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x: number = 42;\n');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    const { promise: rebuildP, resolve: rebuildDone } = Promise.withResolvers<{
+      updates?: { id: string }[];
+    }>();
+    const handle = watch({
+      entryPoints: [join(dir, 'entry.ts')],
+      outfile: join(dir, 'bundle.js'),
+      sourcemap: true,
+      devMode: true,
+      emitDiskSourcemap: false,
+      onReady() {
+        readyDone();
+      },
+      onRebuild(event) {
+        rebuildDone(event);
+      },
+    });
+    await readyP;
+
+    await new Promise((r) => setTimeout(r, 100));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x: number = 7;\n');
+    const event = await rebuildP;
+    expect(event.updates).toBeDefined();
+    expect(event.updates!.length).toBeGreaterThan(0);
+
+    const moduleId = event.updates![0].id;
+    const json = handle.getHmrSourceMap(moduleId);
+    expect(json).not.toBeNull();
+    expect(json).toContain('"version":3');
+
+    expect(handle.getHmrSourceMap('does/not/exist')).toBeNull();
+
+    handle.stop();
+    rmSync(dir, { recursive: true });
+  }, 15000);
+});
