+++ packages/core/test/core/watch/source-maps-rebuilds/hmr-swaps/mixed-getters.ts
@@ -1,0 +1,57 @@
+import {
+  describe,
+  expect,
+  join,
+  mkdtempSync,
+  rmSync,
+  test,
+  tmpdir,
+  watch,
+  writeFileSync,
+} from '../helpers';
+
+describe('watch() > source maps rebuilds > hmr swaps > mixed getters', () => {
+  test('getBundleSourceMap + getHmrSourceMap 교대 호출 — 상호 간섭 없음', async () => {
+    const dir = mkdtempSync(join(tmpdir(), 'zntc-lazy-mix-'));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 42;\n');
+
+    const { promise: readyP, resolve: readyDone } = Promise.withResolvers<void>();
+    const { promise: rebuildP, resolve: rebuildDone } = Promise.withResolvers<{
+      updates?: Array<{ id: string }>;
+    }>();
+    const handle = watch({
+      entryPoints: [join(dir, 'entry.ts')],
+      outfile: join(dir, 'bundle.js'),
+      sourcemap: true,
+      devMode: true,
+      emitDiskSourcemap: false,
+      onReady() {
+        readyDone();
+      },
+      onRebuild(event) {
+        rebuildDone(event);
+      },
+    });
+    await readyP;
+
+    await new Promise((r) => setTimeout(r, 100));
+    writeFileSync(join(dir, 'entry.ts'), 'export const x = 99;\n');
+    const event = await rebuildP;
+    const id = event.updates![0].id;
+
+    for (let i = 0; i < 3; i++) {
+      const bundleJson = handle.getBundleSourceMap();
+      expect(bundleJson).not.toBeNull();
+      expect(JSON.parse(bundleJson!).version).toBe(3);
+
+      const hmrJson = handle.getHmrSourceMap(id);
+      expect(hmrJson).not.toBeNull();
+      const hm = JSON.parse(hmrJson!);
+      expect(hm.version).toBe(3);
+      expect(hm.sources.length).toBe(1);
+    }
+
+    handle.stop();
+    rmSync(dir, { recursive: true });
+  }, 15000);
+});
